experimental/rule_inference/requirements.txt
tree-sitter
tree-sitter-languages
attrs
openai
polyglot-piranha
toml
pytest
flask
flask-socketio
